Initial import. Fix sloppy connection handling.
python-mpd doesn't seem to have a way to recover from dropped
connections. Various exceptions will be raised depending on what is
called; socket.error, AttributeError, and mpd.ConnectionError are the
candidated I've seen so far.

First of all, we're wrapping mpd.MPDClient up so we don't have to deal
with this. It'll try to reconnect when necessary and raises a _single_
exception (mpd.ConnectionError) when it fails. We might eventually want
to poll for a valid connection a la ncmpcpp but we have no ui yet and
there're more important things to do. =P